Move basic_streambuf to streambuf.hpp (API Change):
* Also change allocator element type to char
